feat(consulta): adiciona filtro de pesquisa por data, e ordenação.

+++ src/services/especieService.ts
@@ -1,15 +1,16 @@
 import especieRepository from '../repositories/especieRepository';
+import { BadRequestError, NotFoundError } from '../helpers/ApiError';
 
 class EspecieService {
+  
   async create(data: { nome: string }) {
-
     if (!data.nome || data.nome.trim() === '') {
-      throw new Error('O nome da espécie é obrigatório.');
+      throw new BadRequestError('O nome da espécie é obrigatório.');
     }
 
     const nameExists = await especieRepository.findByName(data.nome);
-    if(nameExists){
-        throw new Error('Espécie já cadastrada.');
+    if (nameExists) {
+      throw new BadRequestError('Espécie já cadastrada.');
     }
 
     return especieRepository.create(data);
@@ -20,14 +21,26 @@ class EspecieService {
   }
 
   async findById(id: number) {
-    return especieRepository.findById(id);
+    const especie = await especieRepository.findById(id);
+    if (!especie) {
+      throw new NotFoundError('Espécie não encontrada.');
+    }
+    return especie;
   }
 
   async update(id: number, data: { nome: string }) {
+    const especie = await especieRepository.findById(id);
+    if (!especie) {
+      throw new NotFoundError('Espécie não encontrada para atualizar.');
+    }
     return especieRepository.update(id, data);
   }
 
   async delete(id: number) {
+    const especie = await especieRepository.findById(id);
+    if (!especie) {
+      throw new NotFoundError('Espécie não encontrada para deletar.');
+    }
     return especieRepository.delete(id);
   }
 }
